+++ server/module-core/src/test/java/com/momo/domain/group/GroupRepositoryTest.java
@@ -1,6 +1,7 @@
-package com.momo.domain.group.domain.repository;
+package com.momo.domain.group;
 
 import com.momo.RepositoryTest;
+import com.momo.domain.group.domain.repository.GroupRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
